+++ PericiasMedicas/person/views.py
@@ -6,8 +6,17 @@ from django.contrib.auth.decorators import login_required
 from django.contrib import messages
 from PericiasMedicas.statescity.models import States2, Cities
 from PericiasMedicas.company.models import Department, Company
+from PericiasMedicas.report.models import Report
 from django.db.models import Q
+from django.db import IntegrityError
 
+#Função para pegar o departamento das pessoas
+def get_department(user):
+    department = Department.objects.filter(profilepersontype__person__cpf=user)    
+    department_id = list()
+    for cc in department:        
+        department_id.append(cc.id) 
+    return department_id
 
 ######  PERSON   #######################################
 @login_required
@@ -52,9 +61,11 @@ def person_detail(request, pk=None, *args, **kwargs):
     return render(request, template_name, context)
 
 @login_required
-def person_edit(request, pk):    
+def person_edit(request, pk):
     template_name='person/form.html'
     data = {}
+    title = "Editar cadastro"
+    data['title'] = title
     ip = request.META['REMOTE_ADDR']
     so = request.META['HTTP_USER_AGENT'].split() # Essa linha mostra o sistema operacional e navegador do usuário   
     person = get_object_or_404(Person, pk=pk)        
@@ -74,16 +85,19 @@ def person_edit(request, pk):
 
 @login_required
 def person_delete(request,pk):
-    person = get_object_or_404(Person, pk=pk)
-    print("request", request.method)
-    if request.method == 'GET':        
-        person.delete()
-        messages.success(request, 'Ação concluída com sucesso.')
-        return redirect('url_people_list')
+    verifica = ProfilePersonType.objects.filter(person_id=pk).exists()
+    if not verifica:
+        person = get_object_or_404(Person, pk=pk)        
+        if request.method == 'GET':        
+            person.delete()
+            messages.success(request, 'Ação concluída com sucesso.')
+            return redirect('url_people_list')
+        else:
+            messages.warning(request, 'Ação não concluída.')
+            return redirect('url_people_list')
     else:
-        messages.warning(request, 'Ação não concluída.')
+        messages.warning(request, 'Não é possível excluir. Pessoa possui dependências.')
         return redirect('url_people_list')
-
 
 #####FIM DO PERSON########################################
 
@@ -110,10 +124,11 @@ def persontype_create(request):
 @login_required
 def persontypes_list(request):
     template_name = "persontype/list.html"
-    persontypes = PersonType.objects.all()    
+    persontypes = PersonType.objects.all() 
+    title = "Perfis Cadastrados"   
     context = {
         'persontypes': persontypes,
-        'title': "Tipos cadastrados"      
+        'title': title,
     }
     return render(request,template_name,context)
 
@@ -144,14 +159,19 @@ def persontype_edit(request, pk):
 
 @login_required
 def persontype_delete(request,pk):
-    persontype = get_object_or_404(PersonType, pk=pk)
-    print("request", request.method)
-    if request.method == 'GET':        
-        persontype.delete()
-        messages.success(request, 'Ação concluída com sucesso.')
-        return redirect('url_persontypes_list')
+    verifica = ProfilePersonType.objects.filter(person_type_id=pk).exists()
+    if not verifica:
+        persontype = get_object_or_404(PersonType, pk=pk)
+        print("request", request.method)
+        if request.method == 'GET':        
+            persontype.delete()
+            messages.success(request, 'Ação concluída com sucesso.')
+            return redirect('url_persontypes_list')
+        else:
+            messages.warning(request, 'Ação não concluída.')
+            return redirect('url_persontypes_list')
     else:
-        messages.warning(request, 'Ação não concluída.')
+        messages.warning(request, 'Não é possível excluir. Existem perfis sendo utilizados.')
         return redirect('url_persontypes_list')
 
 
@@ -184,8 +204,10 @@ def profilepersontype_save(request):
     data['form'] = form
     return render(request,'profilepersontype/form.html',data)
 
+@login_required
 def profilepersontype_create(request,pk):
     template_name = 'profilepersontype/form.html'
+    # Identifica qual o departamento pelo PK
     department = Department.objects.get(pk=pk)
     company = Company.objects.get(pk=department.company_id)
     title1 = "Empresa: "
@@ -193,32 +215,24 @@ def profilepersontype_create(request,pk):
     title3 = "Cadastro de Funcionário"
     data = {
         'department_name': department.name,  
-        'department_id': department.id,       
+        'department': department,       
         'title1': title1,
         'title2': title2,
         'title3': title3,
         'company_name': company.name
     }           
-    if request.method == 'GET':        
-        print("Entrou o GET")
-        form = ProfilePersonTypeForm()
-        data['form'] = form
-        return render(request,template_name,data)
-        
+    if request.method == 'POST':           
+        print("Entrou no POST")
+        form = ProfilePersonTypeForm(request.POST or None, department_id = department.id)
+        print("Passei do form")                      
+        if form.is_valid():                              
+            form.save()     
+            return redirect('url_profilepersontypes_list', pk=pk)# O PK aqui é do Department
+        else:            
+            print("algo não está valido.",form.errors)
     else:
-        print("Entrou no ElSE")
-        form = ProfilePersonTypeForm(request.POST or None)
-        print("Passei do form")
-        data = {}               
-        if form.is_valid():
-            person = form.save(commit=False)
-            department = person.department_id
-            print("Valor do department id", department)
-            person.save()                        
-            return redirect('url_profilepersontypes_list', pk=department)
-        else:
-            print("algo não está valido.")
-            form = ProfilePersonTypeForm()
+        form = ProfilePersonTypeForm(department_id = department.id)
+
     data['form'] = form
     return render(request,'profilepersontype/form.html',data)    
     
@@ -235,7 +249,8 @@ def profilepersontypes_list(request,pk):
         'department_id': department.id,
         'title1': title1,
         'title2': title2,
-        'company_name': company.name
+        'company_name': company.name,
+        'company': company
     }
     if request.method == 'GET':               
         profilepersontypes = ProfilePersonType.objects.filter(department_id=pk) 
@@ -269,7 +284,7 @@ def profilepersontype_edit(request, pk):
     title3 = "Cadastro de Funcionário"
     data = {
         'department_name': department.name,  
-        'department_id': department.id,               
+        'department': department,               
         'title1': title1,
         'title2': title2,
         'title3': title3,
@@ -277,12 +292,14 @@ def profilepersontype_edit(request, pk):
     }                
     user_created = profilepersontype.user_created # Esta linha faz com que o user_created não seja modificado, para mostrar quem criou esta pessoa
     profilepersontype_id = profilepersontype.id
-    form = ProfilePersonTypeForm(request.POST or None, instance=profilepersontype, person_name=profilepersontype.person.name, person_id=profilepersontype.person_id)
+    form = ProfilePersonTypeForm(request.POST or None, instance=profilepersontype, department_id = department.id, person_name=profilepersontype.person.name, person_id=profilepersontype.person_id)
     if form.is_valid():        
         profilepersontype = form.save(commit=False)
         profilepersontype.user_created = user_created               
         profilepersontype.save()
-        return redirect('url_profilepersontype_detail',pk)    
+        return redirect('url_profilepersontype_detail',pk)
+    else:
+        print("algo não está valido.",form.errors)   
     
     data['form'] = form
     return render(request,template_name,data)
@@ -290,33 +307,64 @@ def profilepersontype_edit(request, pk):
 @login_required
 def profilepersontype_delete(request,pk):
     profilepersontype = get_object_or_404(ProfilePersonType, pk=pk)
-    print("request", request.method)
-    if request.method == 'GET':        
-        profilepersontype.delete()
-        messages.success(request, 'Ação concluída com sucesso.')
-        return redirect('url_profilepersontypes_list', profilepersontype.department_id)
+    
+    if request.method == 'GET':   
+        try:             
+            profilepersontype.delete()
+            messages.success(request, 'Ação concluída com sucesso.')
+            return redirect('url_profilepersontypes_list', profilepersontype.department_id)        
+        except IntegrityError:
+            messages.warning(request, 'Não é possível excluir. Existe dependência.')
+            return redirect('url_profilepersontypes_list', profilepersontype.department_id) 
+        
     else:
         messages.warning(request, 'Ação não concluída.')
         return redirect('url_profilepersontypes_list', profilepersontype.department_id)
 
+@login_required
+def show_profile_person_deparment(request):    
+    template_name = 'profilepersontype/show_profile_person_deparment.html'
+    department = Department.objects.filter(id__in = get_department(request.user.username)).order_by('company_id').distinct('company_id')
+    company_id = ""
+    for dp in department:
+        company_id = dp.company_id
+    company = Company.objects.get(pk=company_id)
+    department = Department.objects.filter(company_id=company.id)
+    ppd = ProfilePersonType.objects.filter(department__in=department).order_by('person_id').distinct('person_id')
+    title1 = "Informações da Empresa: "
+    title2 = "Informações dos Departamentos"
+    title3 = "Informações dos funcionários"    
+    data = {
+        'company': company,
+        'department': department,                
+        'profilepersontypes': ppd,
+        'title1': title1,
+        'title2': title2,
+        'title3': title3,        
+    }
+    return render(request,template_name, data)
+    
 
 ######### FIM PROFILE PERSON TYPE #######################
 
 ######### DOCTOR ################
+
+###SOMENTE O PERITO PODERÁ CADASTRAR SEU CURRICULO
 @login_required
 def doctor_create(request):
     template_name = 'doctor/form.html'
     data = {}
-    data['title'] = "Cadastro dos Peritos"        
+    data['title'] = "Cadastro dos Peritos"
+    person = Person.objects.get(cpf=request.user.username)        
     if request.method == 'POST':        
-        form = DoctorForm(request.POST)
+        form = DoctorForm(request.POST,person=person)
         if form.is_valid():
             form.save()                        
             return redirect('url_doctors_list')
         else:
             print("algo não está valido.")
     else:
-        form = DoctorForm()             
+        form = DoctorForm(person=person)             
     
     data['form'] = form
     return render(request,template_name,data)
@@ -324,8 +372,9 @@ def doctor_create(request):
 @login_required
 def doctors_list(request):
     template_name = "doctor/list.html"
-    title = "Especialidades Médicas"
-    doctors = Doctor.objects.all()    
+    title = "Lista de Médicos"
+    #doctors = Doctor.objects.all()    
+    doctors = Doctor.objects.filter(profile_person_type__person__cpf=request.user.username)    
     context = {
         'doctors': doctors,
         'title': title      
@@ -345,9 +394,11 @@ def doctor_detail(request, pk=None, *args, **kwargs):
 def doctor_edit(request, pk):    
     template_name='doctor/form.html'
     data = {} 
-    doctor = get_object_or_404(Doctor, pk=pk)        
+    data['title'] = "Editar informações do Perito"
+    doctor = get_object_or_404(Doctor, pk=pk)
+    person = Person.objects.get(cpf=request.user.username)        
     user_created = doctor.user_created # Esta linha faz com que o user_created não seja modificado, para mostrar quem criou esta pessoa
-    form = DoctorForm(request.POST or None, instance=doctor)
+    form = DoctorForm(request.POST or None, instance=doctor, person=person)
     if form.is_valid():        
         doctor = form.save(commit=False)
         doctor.user_created = user_created               
@@ -359,13 +410,20 @@ def doctor_edit(request, pk):
 
 @login_required
 def doctor_delete(request,pk):
-    doctor = get_object_or_404(Doctor, pk=pk)    
-    if request.method == 'GET':        
-        doctor.delete()
-        messages.success(request, 'Ação concluída com sucesso.')
-        return redirect('url_doctors_list')
+    doctor = get_object_or_404(Doctor, pk=pk) 
+    verifica = Report.objects.filter(profile_person_type__doctor__id=pk, profile_person_type__doctor__medical_specialty_id=doctor.medical_specialty_id)
+    print("verifica",verifica.query)
+    if not verifica:
+        doctor = get_object_or_404(Doctor, pk=pk)        
+        if request.method == 'GET':        
+            doctor.delete()
+            messages.success(request, 'Ação concluída com sucesso.')
+            return redirect('url_doctors_list')
+        else:
+            messages.warning(request, 'Ação não concluída.')
+            return redirect('url_doctors_list')
     else:
-        messages.warning(request, 'Ação não concluída.')
+        messages.warning(request, 'Este perito não pode ser deletado. Ele possui Laudo aprovado ou em produção.')
         return redirect('url_doctors_list')
 
 
@@ -428,14 +486,18 @@ def medicalspecialty_edit(request, pk):
 
 @login_required
 def medicalspecialty_delete(request,pk):
-    medicalspecialty = get_object_or_404(MedicalSpecialty, pk=pk)
-    print("request", request.method)
-    if request.method == 'GET':        
-        medicalspecialty.delete()
-        messages.success(request, 'Ação concluída com sucesso.')
-        return redirect('url_medicalspecialties_list')
+    verifica = Doctor.objects.filter(medical_specialty__id=pk)
+    if not verifica:
+        medicalspecialty = get_object_or_404(MedicalSpecialty, pk=pk)    
+        if request.method == 'GET':        
+            medicalspecialty.delete()
+            messages.success(request, 'Ação concluída com sucesso.')
+            return redirect('url_medicalspecialties_list')
+        else:
+            messages.warning(request, 'Ação não concluída.')
+            return redirect('url_medicalspecialties_list')
     else:
-        messages.warning(request, 'Ação não concluída.')
+        messages.warning(request, 'Existem Peritos com esta especialidades. Não é possível excluir.')
         return redirect('url_medicalspecialties_list')
 
 ######## FIM MedicalSpecialty ###
